Bug Fixing and Question fetch re-structuring
Fixed some bug that prevents the question from navigating to the next
question on multiple selection.

Also, the app can now  displayed the ‘save progress’ at the end of the
final category.

I noticed that there is sufficient time lag when the app is moving to
the next question category, so I re-structure the fetch process to get
all the question at the begining.